+++ demo/js/basicDemo.js
@@ -58,7 +58,7 @@ class DemoOne extends React.Component {
           className="AriaMenuButton"
           onSelection={this.handleSelection.bind(this)}
         >
-          <Button className="AriaMenuButton-trigger">Select a word</Button>
+          <Button tag='button' className="AriaMenuButton-trigger">Select a word</Button>
           <Menu>
             <ul className="AriaMenuButton-menu">{menuItemElements}</ul>
           </Menu>