+++ src/Button.js
@@ -85,12 +85,13 @@ class AriaMenuButtonButton extends React.Component {
       onBlur: this.context.ambManager.handleBlur
     };
 
-    let reserved = checkedProps;
+    const reserved = {};
+    specialAssign(reserved, checkedProps);
     // The disabled property should be passed down to the Button element
     // if the tag has support for disabled attribute. So it needs to be removed
     // from the reserved property object
     if (disabledSupportedTags().indexOf(props.tag) >= 0) {
-      reserved = { children: checkedProps.children, tag: checkedProps.tag };
+      delete reserved.disabled;
     }
     specialAssign(buttonProps, props, reserved);